moved config tests to ts
